requirements.txt
langchain
langchain-community
OpenAI
python-dotenv
huggingface_hub
streamlit